+++ src/test.mjs
@@ -3,6 +3,13 @@ import { isObject } from './isObject'
 import { extractFiles } from './extractFiles'
 import { ReactNativeFile } from './ReactNativeFile'
 
+// eslint-disable-next-line no-console
+console.log(
+  `Testing ${
+    process.execArgv.includes('--experimental-modules') ? 'ESM' : 'CJS'
+  } library with ${process.env.NODE_ENV} NODE_ENV…\n\n`
+)
+
 // eslint-disable-next-line require-jsdoc
 function mock() {
   const file = new ReactNativeFile({ name: '', type: '', uri: '' })